Added some new features and fixed some bugs.
One of the new features is the .current class within the card element now changes the entire background to red with an alpha effect. If the user hovers over the card it will transition back to the normal background color.

Fixed a bug with the accessibility close button taking the dark uni color.

Fixed a bug with some of the elements within the card being overridden by the user.